Add modernization category, move double-precision
