+++ fortitude/src/rules/modernization/double_precision.rs
