+++ fortitude/src/rules/modernization/mod.rs
@@ -1,5 +1,4 @@
-pub mod implicit_kinds;
-pub mod kind_suffixes;
+pub mod double_precision;
 
 #[cfg(test)]
 mod tests {
@@ -15,12 +14,11 @@ mod tests {
     use crate::settings::Settings;
     use crate::test::test_path;
 
-    #[test_case(Rule::NoRealSuffix, Path::new("P001.f90"))]
-    #[test_case(Rule::ImplicitRealKind, Path::new("P021.f90"))]
+    #[test_case(Rule::DoublePrecision, Path::new("MOD001.f90"))]
     fn rules(rule_code: Rule, path: &Path) -> Result<()> {
         let snapshot = format!("{}_{}", rule_code.as_ref(), path.to_string_lossy());
         let diagnostics = test_path(
-            Path::new("precision").join(path).as_path(),
+            Path::new("modernization").join(path).as_path(),
             &[rule_code],
             &Settings::default(),
         )?;
